+++ tests/licenses/entities/actionMessagesGetResult.test.ts
@@ -1,13 +1,21 @@
 import { expect } from 'chai';
-import { ActionMessagesGetResult, LicenseGetFields } from '../../../src';
-import { PAYLOAD_SCHEMA_LICENSE } from '../licensesClient.test';
+import {
+  ActionMessagesGetResult,
+  ActionMessagesGetResultData,
+  ActionMessagesGetResultFields,
+} from '../../../src';
 
 describe('ActionMessagesGetResult', () => {
-  const result = new ActionMessagesGetResult(
-    PAYLOAD_SCHEMA_LICENSE.data.license[
-      LicenseGetFields.COLUMN_ACTION_MESSAGES
-    ][0],
-  );
+  const ACTION_MESSAGE_PAYLOAD: ActionMessagesGetResultData = {
+    [ActionMessagesGetResultFields.COLUMN_ACTION]: 'action',
+    [ActionMessagesGetResultFields.COLUMN_MESSAGE]: 'message',
+    [ActionMessagesGetResultFields.COLUMN_MAX_DECREASE]: 5,
+    [ActionMessagesGetResultFields.COLUMN_SUPPORTED_UNTIL]: 'supported_until',
+    [ActionMessagesGetResultFields.COLUMN_SUSPEND_DATE]: 'suspend_date',
+  };
+
+  const result = new ActionMessagesGetResult(ACTION_MESSAGE_PAYLOAD);
+
   describe('constructor', () => {
     it('sets a default value for object', function () {
       expect(result).to.be.instanceof(ActionMessagesGetResult);
@@ -16,11 +24,7 @@ describe('ActionMessagesGetResult', () => {
 
   describe('toJSON', () => {
     it('sets a default value for toJSON function', function () {
-      expect(result.toJSON()).to.eqls(
-        PAYLOAD_SCHEMA_LICENSE.data.license[
-          LicenseGetFields.COLUMN_ACTION_MESSAGES
-        ][0],
-      );
+      expect(result.toJSON()).to.eqls(ACTION_MESSAGE_PAYLOAD);
     });
   });
 });