+++ tests/licenses/licensesClient.test.ts
@@ -261,7 +261,7 @@ export const PAYLOAD_SCHEMA_LICENSE: GetData<LicenseGet> = {
         [ActiveSeatsFindResultFields.COLUMN_LAST_UPDATE]: new Date().toISOString(),
       },
       [LicenseGetFields.COLUMN_ACTIVATION_DATETIME]: 'activation_datetime',
-      [LicenseGetFields.COLUMN_EXPIRY_DATETIME]: 'expery_datetime',
+      [LicenseGetFields.COLUMN_EXPIRY_DATETIME]: 'expiry_datetime',
       [LicenseGetFields.COLUMN_AUTO_RENEW]: true,
       [LicenseGetFields.COLUMN_MESSAGE]: 'message',
       [LicenseGetFields.COLUMN_ACTIONS]: {
@@ -309,6 +309,52 @@ export const PAYLOAD_SCHEMA_LICENSE: GetData<LicenseGet> = {
         },
       },
       [LicenseGetFields.COLUMN_ARROW_SUB_CATEGORIES]: ['string'],
+    },
+  },
+};
+
+export const PAYLOAD_SCHEMA_LICENSE_WITHOUT_OPTIONAL_FIELDS: GetData<LicenseGet> = {
+  status: 200,
+  data: {
+    license: {
+      [LicenseGetFields.COLUMN_LICENSE_ID]: '123456',
+      [LicenseGetFields.COLUMN_PARENT_LICENSE_ID]: 'parent_license_id',
+      [LicenseGetFields.COLUMN_FRIENDLY_NAME]: 'friendly_name',
+      [LicenseGetFields.COLUMN_CUSTOMER_REF]: 'customer_name',
+      [LicenseGetFields.COLUMN_STATE]: 'state',
+      [LicenseGetFields.COLUMN_SERVICE_REF]: 'service_ref',
+      [LicenseGetFields.COLUMN_SKU]: 'sku',
+      [LicenseGetFields.COLUMN_NAME]: 'name',
+      [LicenseGetFields.COLUMN_SEATS]: 10,
+      [LicenseGetFields.COLUMN_ACTIVE_SEATS]: {
+        [ActiveSeatsFindResultFields.COLUMN_NUMBER]: 3,
+        [ActiveSeatsFindResultFields.COLUMN_LAST_UPDATE]: new Date().toISOString(),
+      },
+      [LicenseGetFields.COLUMN_ACTIVATION_DATETIME]: 'activation_datetime',
+      [LicenseGetFields.COLUMN_EXPIRY_DATETIME]: 'expiry_datetime',
+      [LicenseGetFields.COLUMN_MESSAGE]: 'message',
+      [LicenseGetFields.COLUMN_ORDER_REFERENCE]: 'reference',
+      [LicenseGetFields.COLUMN_ORDER]: {
+        [OrderGetFields.COLUMN_LINK]: 'link',
+        [OrderGetFields.COLUMN_REFERENCE]: 'reference',
+      },
+      [LicenseGetFields.COLUMN_VENDOR_LICENSE_ID]: 'vendor_license_id',
+      [LicenseGetFields.COLUMN_PERIODICITY]: 'periodicity',
+      [LicenseGetFields.COLUMN_TERM]: 'term',
+      [LicenseGetFields.COLUMN_CATEGORY]: 'category',
+      [LicenseGetFields.COLUMN_PROGRAM]: 'program',
+      [LicenseGetFields.COLUMN_ASSOCIATED_SUBSCRIPTION_PROGRAM]:
+        'associated_subscription_program',
+      [LicenseGetFields.COLUMN_PRICE]: {
+        [LicensePriceGetFields.COLUMN_UNIT]: {
+          [BuySellFields.COLUMN_BUY]: 2,
+          [BuySellFields.COLUMN_SELL]: 3,
+        },
+        [LicensePriceGetFields.COLUMN_TOTAL]: {
+          [BuySellFields.COLUMN_BUY]: 2,
+          [BuySellFields.COLUMN_SELL]: 3,
+        },
+      },
     },
   },
 };
@@ -771,6 +817,17 @@ describe('LicensesClient', () => {
         '123456',
       );
       expect(result).to.eql(PAYLOAD_SCHEMA_LICENSE);
+    });
+
+    it('calls the get method and returns its result without optional fields', async () => {
+      nock(LICENSES_MOCK_URL)
+        .get(LICENSE_MOCK_URL_GET_LICENSE)
+        .reply(200, PAYLOAD_SCHEMA_LICENSE_WITHOUT_OPTIONAL_FIELDS);
+
+      const result: GetData<LicenseGet> = await getLicenseClient.getLicense(
+        '123456',
+      );
+      expect(result).to.eql(PAYLOAD_SCHEMA_LICENSE_WITHOUT_OPTIONAL_FIELDS);
     });
   });
 });